requirements.txt
altgraph==0.17
asgiref==3.4.1
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.2.5
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
macholib==1.14
mccabe==0.6.1
norminette==3.3.2
Pillow==8.3.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pylint==2.6.0
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
wrapt==1.12.1
